Add files via upload
Please start working on the project with given data set(titanic.zip). You all need to use SSIS and MS SQL server for the Project.
You project should include following answer (but not limited) from the dataset.



1. Load the Data Using the Bulk Load
2. Archived the data after load zipping it in archived folder.
3. Prepare the final data that can answer following Question:
a. Total no of survival
a.a Gender wise
a.b Age group wise
a.c According to Fare paid in range.